+++ install.sh
@@ -1,4 +1,8 @@
 #!/bin/sh
+# SPDX-FileCopyrightText: 2025 Ryan Cao <hello@ryanccn.dev>
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+
 set -euf
 
 ansi_red=""