+++ src/cli/complete.rs
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: 2025 Ryan Cao <hello@ryanccn.dev>
+//
+// SPDX-License-Identifier: GPL-3.0-or-later
+
 use clap_complete::CompletionCandidate;
 use color_eyre::eyre::Result;