+++ src/cli/env_file.rs
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: 2025 Ryan Cao <hello@ryanccn.dev>
+//
+// SPDX-License-Identifier: GPL-3.0-or-later
+
 use dotenvy::Result;
 
 #[derive(Clone, Debug)]