+++ src/package_json.rs
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: 2025 Ryan Cao <hello@ryanccn.dev>
+//
+// SPDX-License-Identifier: GPL-3.0-or-later
+
 use std::{fs, path::Path};
 
 use owo_colors::{OwoColorize as _, Stream};